+++ src/main/java/dto/responce/ExchangeRateRespDTO.java
@@ -1,8 +1,10 @@
 package dto.responce;
 
+import java.math.BigDecimal;
+
 public record ExchangeRateRespDTO(
         int id,
         CurrencyRespDTO baseCurrency,
         CurrencyRespDTO targetCurrency,
-        double rate) {
+        BigDecimal rate) {
 }
